backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
numpy==1.24.3
faiss-cpu==1.7.4
requests==2.31.0
pdfplumber==0.9.0
pytesseract==0.3.10
python-docx==0.8.11
Pillow==10.0.0
sentence-transformers==2.2.2
Werkzeug==2.3.7
python-dotenv==1.0.0